move Legend component
