+++ client/components/DonutChart/styles.scss
@@ -11,14 +11,3 @@
 .mainSvg {
   filter: drop-shadow( 2px 2px 4px rgba(0,0,0,.5) );
 }
-
-.legend {
-  font-size: 1.2rem;
-  font-weight: bold;
-}
-
-.value {
-  font-size: 1rem;
-  color: $gray;
-  padding-left: 0.5em;
-}
