+++ client/components/Legend/LegendItem.js
