+++ client/components/Legend/index.js
