+++ client/components/Legend/styles.scss
@@ -1,0 +1,12 @@
+@import 'theme/variables';
+
+.legend {
+  font-size: 1.2rem;
+  font-weight: bold;
+}
+
+.value {
+  font-size: 1rem;
+  color: $gray;
+  padding-left: 0.5em;
+}
